// src/app/models/models.ts
export interface LoggedUserDto {
    id: string | null;
    userName: string | null;
    learningPath: string | null;
}
// ...
export interface UpdateUserLearningPathCommand {
    userId: string | null;
    learningPathId: string | null;
}

export interface UserDto {
    id: string | null;
    email: string | null;
    userName: string | null;
    learningPathId: string | null;
    rolesIds: string[] | null;
    status: string | null;
    isBlocked: boolean;
    invitedBy: string | null;
    creationDate: Date;
}

export interface LearningPathForUserDto {
    readonly id: string | null;
    readonly name: string | null;
}

export interface RoleForUserDto {
    readonly id: string | null;
    readonly name: string | null;
    readonly description: string | null;
}

export interface UserViewModel {
    user: UserDto;
    learningPaths: LearningPathForUserDto[] | null;
    roles: RoleForUserDto[] | null;
}

export interface UpdateUserRolesCommand {
    userId: string | null;
    rolesIds: string[] | null;
}

export interface BlockUserCommand {
    userId: string | null;
}

export interface UnblockUserCommand {
    userId: string | null;
}